[#23] Refactoring
