+++ src/baluster/__init__.py
@@ -1,1 +1,2 @@
-from .base import Holder, MultipleExceptions, ContextManagerReusedError  # noqa
+from .base import Holder                                               # noqa
+from .exceptions import MultipleExceptions, ContextManagerReusedError  # noqa
