+++ src/baluster/manager.py
@@ -1,0 +1,24 @@
+from .exceptions import ContextManagerReusedError
+
+
+class Manager:
+
+    def __init__(self, managed):
+        self._managed = managed
+        self._active = False
+
+    def __enter__(self):
+        if self._active is True:
+            raise ContextManagerReusedError()
+        self._active = True
+        return self._managed
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self._active = False
+        self._managed.close()
+
+    async def __aenter__(self):
+        return self._managed
+
+    async def __aexit__(self, exc_type, exc_value, traceback):
+        await self._managed.aclose()
