+++ src/baluster/utils.py
@@ -1,0 +1,4 @@
+def make_if_none(obj, default):
+    if obj is not None:
+        return obj
+    return default
